Committing agent exploration demo.
In this demo:
The service namespace used by the plugin is "gpigf", it provides a service
called "processAgentPath".

The js uses this service to remove the area an agent has explored from a
target area polygon. In the js there is also an alternative implementation
(mainly for education) that uses existing services to achieve the same thing
(see the function called agentPath2 in AgentDemo.js)

Using the web UI, create a target area polygon, then create an agents path
with the other tool. The java plugin will then calculate the shape of the
new target area (i.e. the area remaining to explore).


Additional demos should be added alongside app.js, the user then changes
the requires line in app.js to load the demo they are interested in.

Additional services, for additional demos, should be added to file in
META-INF/services (I think!)