An example of how to shrink an area. Growing/Shrinking is different from scaling.
There is no java plugin for this, it uses existing services.
In the demo, draw a shape then use the other tools to grow and shrink the
shape by clicking on it.

Note the change made in app.js to use this demo. Change it back for the
agent demo.