Huge commit of 'pres' folders.
In these folders are a proposed framework for our actual presentation.
Whether we use this or not is another thing. This work was still needed
so that we had a clearer idea of our architecture. Have a browse, the js
is heavily commented, but I do have atrocious grammar.

The gpigf-pres folder contains the java for each plugin - as individual
projects. The webapp-pres folder contains the js implementation for the
presentation.